Rename fragment as CodeEditorAutocomplete
