+++ src/fragments/CodeEditorAutocomplete.tsx
